Attempt to resolve for loop variable issue (WIP)
The problem is:
> for i in range(5):
>     x = i
Python assigns x=4, stella x=5 because the test is run after the
increment.

This commit adds a second test which comes before the increment. The
jump at the end then goes directly to the body because the test has
already run.

However, the tests crash now with a segfault.